(streams): An attempt to understand data streaming
I started by writing requirements. When they were complete, I
started with implementation. After a while, I tried to use one
stream to read a file and two writers to move it. I wasn't sure
how to do it, so I tried to pipe the calls, but it didn't work.
I started with the file 'how-streams-transfer-data'. When I got
an error, I started working on 'one-source-multiple-destinations'
to learn how to fix it.

Contents:
- C++ sample program
- how-streams-transfer-data
  (transfers the C++ executable in binary and txt format)
  - binary format (if runnable, data transfer worked)
  - txt format (i wanted to read contents of binary file)
- one-source-multiple-destinationswq