Monitor TCP traffic on a port (#40)
The TCP protocol is a set of rules on how the traffic gets sent
over the internet. I wanted to inspect those packages a few months
back. However, I didn't know how and failed. I realized today that
traffic gets sent through the port. I verified it with a quick
google search. I set up a sample that creates a server, a script
that monitors TCP traffic on a port, and a client that requests
data.

A non-encrypted server makes TCP packages readable. A client sends
a request to my server, so there is something to read. Lastly, a
script that monitors exchanged traffic.