feat(runner): add dangerous case
Remember to handle the dangerous case where we may overwrite or
delete configuration. We should take all measures of precaution
to prevent such behavior. It is a disaster if any of our operations
fails and overwrites the configuration. If that is the only config
the user saved, this case is even worse.