feat: add sample directory
Provide a small sample to improve focus. The small directory helps
emphasize the goal. This sample provides a base structure for CI. I
wanted to reference it in a blog post. The readers can clearly see
what I did. The repository is way too big to use as a reference in
a blog post.