move example to docs
The example doesn't have any source code. We use docs for such code.
I don't like that it has file with it, but I will place it in
directory and use README.md to achieve the same effect. When we
click on the directory, it will read the README.md