implement min heap
* Copy the original implementation from Codepen

I like to use this repository for historical purposes, so I store
everything I've done here. I want to use it as a tool to look back and
reflect on stuff that I've worked on. I like when I see things that I've
tried and compare my current situation to that at the time of code I
worked on.

* Use camelCase instead of underscores and lowercase

I didn't follow convention of other files in this repository so I used
camelCase because it is preffered in JavaScript community. However,
other files in this directory follow lowercase and underscore
convention. I didn't feel like refactoring because it would take time
and it didn't feel fun any more.