add readme file
Save what is the core idea of OpenGL because now I feel like I finally
"GET" it. This is what I was looking for, this wording. Basically it
matches everything I've read so far. Data is saved inside OpenGL context
and we use instructions to execute various OpenGL mechanics to operate
on that state.