add thoughts on understanding beginner mindset
I was browsing git commit history and found a commit when I was setting
up CPP experiments directory. I remembered how lost and disoriented I
felt. That made me realize how this repository helps me go back in time
and remember how it feels to be new and unfamiliar. What were the
problems that I used to have. I had to document these thoughts because I
feel like they are so important to remember in the future.