docs(cpp): save thoughts on repository's beginning (#162)
* Add thoughts on understanding beginner mindset

I was browsing git commit history and found a commit when I was setting
up CPP experiments directory. I remembered how lost and disoriented I
felt. That made me realize how this repository helps me go back in time
and remember how it feels to be new and unfamiliar. What were the
problems that I used to have. I had to document these thoughts because I
feel like they are so important to remember in the future.

* Rename file names to use underscore

The majority of experiments use underscore instead of slash. Since I
modified docs to add a story I refactored the files to be consistent in
the whole directory.