feat(cpp): link glew and sdl libraries (#165)
After setting up Glew and SDL C++ libraries using Visual Studio, I had
a sudden idea of how to set it up manually using compiler and batch
script. That process was fun and helped me see how tedious setting up
a C++ toolchain can be.

I am not very happy with distributing `.dll` and `.lib` files because
they are generated files, but I wanted something quick that would work
for anyone on windows 64-bit architecture. An alternative was to set up
a location somewhere and use a batch script to download SDL and GLEW
library, then position them to their location. However, using code to
do that would be error and bug-prone, and I didn't see a worthy
downside of committing the binaries.