rename batch file and add documentation
The batch file used to be called compile.bat but it compiles and runs
the C++ file. However since it compiles and runs it is more suitable to
call it run.bat because that reflects what it does.

* Add documentation

I don't want to read the code of the batch file to be able to use it.