feat(cpp): play with operator overloading (#172)
I had a struct edge and wanted to create a priority queue of edges for
Kruskal's algorithm. It uses an operator less than to compare edges. It
wasn't working, and I didn't know why, so it took me a while to figure
it out.

I wanted to get comfortable with the solution to the error on why the
standard library algorithm didn't accept the struct that I created.

I didn't know that error was caused because I overloaded the operator
as a non-const function which may mutate the object.