feat(cpp): separate declaration and definition (#173)
So far, I didn't think of separating the two. I always created header
and source files because resources say it is the way to go. This sample
shows what happens after the preprocessor goes through includes in
source files.

I wanted to try out the concept of splitting declaration and
implementation manually. It is right before me, but it is not apparent
to a beginner. It was in front of me all the time in the form of header
and cpp files, but it became apparent only once I tried to split the
declaration and definition by myself.

It wasn't apparent to me that this is how header and cpp files work.