chore: add general documentation directory (#178)
General documentation is now in sk-experiments/docs/ directory.

General documentation was scattered around in language directories and
it didn't belong there. For example, C++ experiment documentation
should be in the directory for C++.

* Move documentation from root README to general documentation

Too much text in the root README.md is terrifying. I moved it to the general
to be more visually pleasing.

* Use root docs directory for general documentation

I noticed that I added documentation to a language's directory. That
isn't correct because general observations and information that's
unrelated to a programming language shouldn't be in a programming language
directory. The solution was to create a general directory for
documentation and the best place was the root directory.

* Delete general documentation from the language docs directory

I moved the general documentation to sk-experiments/docs/ so i deleted
it from the previous place cpp/docs.

* Use dash instead of underscore

The sk-experiments directory uses dashes so I'll stay consistent and use
them in sk-experiments/docs/ as well. The directory
sk-experiments/cpp/docs/ uses underscore so the documentation there
use underscores. Since the documentation is, in general same for
anything I do I should stay consistent and use the same convention.
Therefore I will use dashes.

* Use dashes in cpp/docs

Every directory uses dashes in the documentation directory to be
consistent. Rename sk-experiments/cpp/docs/ documentation files to use
dashes instead of underscores.