arrange labels vertically
I used div inside label but label is inline element while div is block
element. Block elements are allowed to be rendered inside inline
elements but it's not correct HTML. Since HTML validation is incredibly
flaky and I discovered this late, I used inline styling to fix it.
Inline styling isn't good either but this project should't grow so it
should be fine.