use boost version 1.79.0
After debugging around a week ago I found out that find_Boost.cmake does
not search for version of boost above 1.79.0 and I found that out
through using flag that allows boost to output what it's searching for.
The simplest solution I found is to downgrade the required version as I
don't have the need to get the latest one.