change script to download dependencies as well
We previously required to create directory and download mongoose
manually. I found out that I could display the content of a file on
GitHub in raw format which allowed me to download it using
Invoke-Webrequest. This change makes it easier for me in the future to
work on this project.