chore: add documentation on how to setup dependencies (#222)
It wasn't clear before how to setup dependencies and what are the set of
steps to get the example to run. I tested the script on another machine
and it worked nicely. Dependencies were set up properly and the example
compiled and ran nicely. I will definitely forget how to run the code in
the future so I added documentation on how to do it.